requirements.txt
rpaframework-windows>=7.3.2
pytest>=7.4.0